Gemini-historical-artifact-description/requirements.txt
pip install streamlit google-generativeai python-dotenv
GOOGLE_API_KEY=your_actual_api_key_here
